+++ credential-verification-service/src/api/verify.rs
@@ -1,6 +1,6 @@
 //! Handler for the verification endpoints.
 
-use crate::api::util::QueryErrorExt;
+use crate::api::util;
 use crate::node_client::NodeClient;
 use crate::types::AppJson;
 use crate::{
@@ -9,8 +9,6 @@ use crate::{
 };
 use anyhow::Context;
 use axum::{Json, extract::State};
-use concordium_rust_sdk::base::hashes::TransactionHash;
-use concordium_rust_sdk::base::transactions::{BlockItem, ExactSizeTransactionSigner, send};
 use concordium_rust_sdk::base::web3id::v1::anchor::{
     CredentialValidityType, VerifiablePresentationV1, VerificationMaterial,
     VerificationMaterialWithValidity, VerificationRequest, VerificationRequestAnchor,
@@ -21,27 +19,18 @@ use concordium_rust_sdk::base::web3id::v1::{
     IdentityCredentialVerificationMaterial,
 };
 use concordium_rust_sdk::common::cbor;
-
-use concordium_rust_sdk::id::types::{AccountCredentialWithoutProofs, ArInfos};
-use concordium_rust_sdk::types::{
-    AccountTransactionEffects, BlockItemSummaryDetails, RegisteredData,
-};
-
 use concordium_rust_sdk::id::types;
+use concordium_rust_sdk::id::types::{AccountCredentialWithoutProofs, ArInfos};
+use concordium_rust_sdk::types::{AccountTransactionEffects, BlockItemSummaryDetails};
 use concordium_rust_sdk::v2::BlockIdentifier;
-use concordium_rust_sdk::web3id::v1::CreateAnchorError;
 use concordium_rust_sdk::{
     base::web3id::v1::anchor::{
         self, PresentationVerificationResult, VerificationAuditRecord, VerificationContext,
     },
-    common::types::TransactionTime,
-    types::WalletAccount,
-    web3id::v1::{
-        AnchorTransactionMetadata, AuditRecordArgument, PresentationVerificationData, VerifyError,
-    },
+    web3id::v1::VerifyError,
 };
 use futures_util::future;
-use std::collections::{BTreeMap, HashMap};
+use std::collections::BTreeMap;
 use std::sync::Arc;
 
 /// Verify Presentation endpoint handler.
@@ -57,7 +46,6 @@ pub async fn verify_presentation(
     let mut client = state.node_client.clone();
 
     // Verify the presentation with respect to the verification request anchor.
-    // note: we do not lock for the account nonce until the anchor submission
     let presentation_verification_result = verify_presentation_with_request_anchor(
         &mut *client,
         block_identifier,
@@ -66,99 +54,43 @@ pub async fn verify_presentation(
     )
     .await?;
 
-    // Lock for account nonce now before we call to create and submit the audit anchor
-    let mut account_sequence_number = state.nonce.lock().await;
+    // Create the audit record
+    let verification_audit_record = VerificationAuditRecord::new(
+        verify_presentation_request.audit_record_id,
+        verify_presentation_request.verification_request,
+        verify_presentation_request.presentation,
+    );
 
-    // create and submit the audit anchor on chain
-    let presentation_verification_data_result = create_and_submit_audit_anchor(
-        &mut *client,
-        &verify_presentation_request,
-        &state,
-        presentation_verification_result,
-        *account_sequence_number,
-    )
-    .await;
+    // Submit the audit anchor if verification was successful
+    let anchor_transaction_hash = if presentation_verification_result.is_success() {
+        let audit_record_anchor =
+            verification_audit_record.to_anchor(verify_presentation_request.public_info);
+        let anchor_data = util::anchor_to_registered_data(&audit_record_anchor)
+            .context("create audit record anchor registration data")?;
 
-    // Check now if the presentation verification was ok
-    match presentation_verification_data_result {
-        Ok(presentation_verification_data) => {
-            let result = match presentation_verification_data.verification_result {
-                PresentationVerificationResult::Verified => VerificationResult::Verified,
-                PresentationVerificationResult::Failed(e) => {
-                    VerificationResult::Failed(e.to_string())
-                }
-            };
-
-            let verify_presentation_response = VerifyPresentationResponse {
-                result,
-                anchor_transaction_hash: presentation_verification_data.anchor_transaction_hash,
-                verification_audit_record: presentation_verification_data.audit_record,
-            };
-
-            // increment states nonce now
-            if presentation_verification_data
-                .anchor_transaction_hash
-                .is_some()
-            {
-                *account_sequence_number = account_sequence_number.next();
-            }
-            Ok(Json(verify_presentation_response))
-        }
-        Err(CreateAnchorError::Query(err)) if err.is_account_sequence_number_error() => {
-            tracing::warn!(
-                "Unable to submit transaction on-chain successfully due to account nonce mismatch. Account nonce will be refreshed and transaction will be re-submitted: {}",
-                err
-            );
-
-            let mut client = state.node_client.clone();
-
-            // Refresh nonce
-            let nonce = client
-                .get_next_account_sequence_number(&state.account_keys.address)
-                .await
-                .context("get next account sequence number")?;
-
-            // resubmit the audit anchor with the updated account sequence number
-            *account_sequence_number = nonce;
-            tracing::info!("Refreshed account nonce successfully.");
-            let presentation_verification_data_result = create_and_submit_audit_anchor(
-                &mut *client,
-                &verify_presentation_request,
-                &state,
-                presentation_verification_result,
-                *account_sequence_number,
-            )
+        let txn_hash = state
+            .transaction_submitter
+            .submit_register_data_txn(anchor_data)
             .await
-            .context("submit audit anchor")?;
+            .context("submit audit anchor register data transaction")?;
 
-            let result = match presentation_verification_data_result.verification_result {
-                PresentationVerificationResult::Verified => VerificationResult::Verified,
-                PresentationVerificationResult::Failed(e) => {
-                    VerificationResult::Failed(e.to_string())
-                }
-            };
+        Some(txn_hash)
+    } else {
+        None
+    };
 
-            let verify_presentation_response = VerifyPresentationResponse {
-                result,
-                anchor_transaction_hash: presentation_verification_data_result
-                    .anchor_transaction_hash,
-                verification_audit_record: presentation_verification_data_result.audit_record,
-            };
+    let result = match presentation_verification_result {
+        PresentationVerificationResult::Verified => VerificationResult::Verified,
+        PresentationVerificationResult::Failed(e) => VerificationResult::Failed(e.to_string()),
+    };
 
-            // finally increase the nonce in the state
-            if presentation_verification_data_result
-                .anchor_transaction_hash
-                .is_some()
-            {
-                *account_sequence_number = account_sequence_number.next();
-            }
+    let verify_presentation_response = VerifyPresentationResponse {
+        result,
+        anchor_transaction_hash,
+        verification_audit_record,
+    };
 
-            Ok(Json(verify_presentation_response))
-        }
-        Err(err) => Err(anyhow::Error::from(err)
-            .context("create and submit request anchor")
-            .into()),
-    }
+    Ok(Json(verify_presentation_response))
 }
 
 /// Perform the full verification of the presentation with respect to the Verification Request Anchor.
@@ -200,15 +132,14 @@ async fn verify_presentation_with_request_anchor(
     };
 
     // Verify Presentation with respect to the Verification Request Anchor
-    let presentation_verification_result: PresentationVerificationResult =
-        anchor::verify_presentation_with_request_anchor(
-            &global_context,
-            &verification_context,
-            &verify_presentation_request.verification_request,
-            &verify_presentation_request.presentation,
-            &request_anchor,
-            &verification_material,
-        );
+    let presentation_verification_result = anchor::verify_presentation_with_request_anchor(
+        &global_context,
+        &verification_context,
+        &verify_presentation_request.verification_request,
+        &verify_presentation_request.presentation,
+        &request_anchor,
+        &verification_material,
+    );
 
     Ok(presentation_verification_result)
 }
@@ -346,92 +277,4 @@ async fn lookup_verification_material_and_validity(
             }
         }
     })
-}
-
-/// Creates and submits the Verification Audit anchor to the chain.
-/// Only submits the anchor if the presentation verification result was a success.
-async fn create_and_submit_audit_anchor(
-    client: &mut dyn NodeClient,
-    verify_presentation_request: &VerifyPresentationRequest,
-    state: &State<Arc<Service>>,
-    verification_result: PresentationVerificationResult,
-    account_sequence_number: concordium_rust_sdk::types::Nonce,
-) -> Result<PresentationVerificationData, CreateAnchorError> {
-    // Prepare Data for Audit anchor submission
-    let verify_presentation_request_clone = verify_presentation_request.clone();
-    let audit_record = VerificationAuditRecord::new(
-        verify_presentation_request_clone.audit_record_id,
-        verify_presentation_request_clone.verification_request,
-        verify_presentation_request_clone.presentation,
-    );
-
-    // submit the audit anchor transaction
-    let anchor_transaction_hash = if verification_result.is_success() {
-        let audit_record_argument =
-            build_audit_record(state, verify_presentation_request, account_sequence_number).await;
-
-        let txn_hash = submit_verification_audit_record_anchor(
-            client,
-            audit_record_argument.audit_record_anchor_transaction_metadata,
-            &audit_record,
-            audit_record_argument.public_info,
-        )
-        .await?;
-        Some(txn_hash)
-    } else {
-        None
-    };
-
-    Ok(PresentationVerificationData {
-        verification_result,
-        audit_record,
-        anchor_transaction_hash,
-    })
-}
-
-async fn submit_verification_audit_record_anchor<S: ExactSizeTransactionSigner>(
-    client: &mut dyn NodeClient,
-    anchor_transaction_metadata: AnchorTransactionMetadata<S>,
-    verification_audit_record: &VerificationAuditRecord,
-    public_info: Option<HashMap<String, cbor::value::Value>>,
-) -> Result<TransactionHash, CreateAnchorError> {
-    let verification_audit_anchor = verification_audit_record.to_anchor(public_info);
-    let cbor = cbor::cbor_encode(&verification_audit_anchor)?;
-    let register_data = RegisteredData::try_from(cbor)?;
-
-    let tx = send::register_data(
-        &anchor_transaction_metadata.signer,
-        anchor_transaction_metadata.sender,
-        anchor_transaction_metadata.account_sequence_number,
-        anchor_transaction_metadata.expiry,
-        register_data,
-    );
-    let item = BlockItem::AccountTransaction(tx);
-
-    // Submit the transaction to the chain.
-    let transaction_hash = client.send_block_item(&item).await?;
-
-    Ok(transaction_hash)
-}
-
-/// Helper function to build the Audit record Argument that will be used in the verify presentation call
-async fn build_audit_record<'s>(
-    state: &'s State<Arc<Service>>,
-    verify_presentation_request: &VerifyPresentationRequest,
-    account_sequence_number: concordium_rust_sdk::types::Nonce,
-) -> AuditRecordArgument<&'s Arc<WalletAccount>> {
-    let expiry = TransactionTime::seconds_after(state.transaction_expiry_secs);
-
-    let audit_record_anchor_transaction_metadata = AnchorTransactionMetadata {
-        signer: &state.account_keys,
-        sender: state.account_keys.address,
-        account_sequence_number,
-        expiry,
-    };
-
-    AuditRecordArgument {
-        audit_record_id: verify_presentation_request.audit_record_id.clone(),
-        public_info: verify_presentation_request.public_info.clone(),
-        audit_record_anchor_transaction_metadata,
-    }
 }