+++ credential-verification-service/src/lib.rs
@@ -6,3 +6,4 @@ pub mod service;
 pub mod types;
 
 mod api;
+mod txn_submitter;