+++ credential-verification-service/src/types.rs
@@ -3,28 +3,22 @@ use axum::extract::FromRequest;
 use axum::extract::rejection::JsonRejection;
 use axum::http::StatusCode;
 use axum::response::{IntoResponse, Response};
-use concordium_rust_sdk::{
-    types::{Nonce, WalletAccount},
-    web3id::did::Network,
-};
+use concordium_rust_sdk::web3id::did::Network;
 use std::fmt::Display;
-use std::sync::Arc;
-use tokio::sync::Mutex;
+
+use crate::txn_submitter::TransactionSubmitter;
 
 /// Holds the service state in memory.
 ///
 /// Note: A new instance of this struct is created whenever the service restarts.
+#[derive(Debug, Clone)]
 pub struct Service {
     /// The client to interact with the node.
     pub node_client: Box<dyn NodeClient>,
     /// The network of the connected node.  
     pub network: Network,
-    /// The key and address of the account submitting the anchor transactions on-chain.
-    pub account_keys: Arc<WalletAccount>,
-    /// The current nonce of the account submitting the anchor transactions on-chain.
-    pub nonce: Arc<Mutex<Nonce>>,
-    /// The number of seconds in the future when the anchor transactions should expiry.  
-    pub transaction_expiry_secs: u32,
+    /// Submitter for transactions
+    pub transaction_submitter: TransactionSubmitter,
 }
 
 /// Extractor with build in error handling. Like [axum::Json](Json) but will use [`RejectionError`] for rejection errors
@@ -49,7 +43,7 @@ pub enum ServerError {
 /// See <https://docs.rs/axum/latest/axum/extract/index.html#customizing-extractor-responses>
 #[derive(Debug, thiserror::Error)]
 pub enum RejectionError {
-    #[error("invalid json in request")]
+    #[error("invalid json in request: {0}")]
     JsonRejection(#[from] JsonRejection),
 }