+++ credential-verification-service/CHANGELOG.md
@@ -1,7 +1,11 @@
 ## Unreleased
 
+- Refactored transaction submit logic
+ 
 ## 0.1.0
 
-- Added account nonce management to the service.
-- Added the logic of the `/verifiable-presentations/create-verification-request` api endpoint flow. This endpoint submits the `verification-request-anchor (VRA)` on-chain.
-- Initial service.
+- Initial service which implements verifying Concordium V1 verifiable presentations and creating verification
+  requests. This includes submitting the verification request anchors (VRA), verifying the request anchor and
+  submitting the verification audit anchor (VAA). The following endpoints exist in the initial version
+  - `/verifiable-presentations/verify`
+  - `/verifiable-presentations/create-verification-request`